This container component fetches information from the hash URL and then from the database in order to render the proper piece of the Foundation.
If props are passed to this component, it ignores the URL hash and uses these instead to fetch database data. Props will be passed to this component when a user clicks on a piece of the Foundation while reading a take. 

The URL functionality of the component allows for easy sharing of the current application state.