Updated the data struture:
Document type just contains the string representing the Document type, i.e. CONSTITUTION.

DocumentExcerpt type is now reprented by a range that is a subset of the Document type. Removed highlightedRange and viewRange which are for the UI.

TakeDocuments now contain DocumentBlocks that contain DocumentExcerpts with the added ranges highlightedRange and viewRange (yet to be implemented).