Created a getNodesInRange function which returns a subset of a FoundationNode array when given a range.
Adjusted getNodeArray to take a string representing the ID of the excerpt to find. It does a DB lookup for the excerpt information based on this ID and returns the proper nodes.

Added a slugify function that turns a string into a valid URL slug.

Added a getExcerpt function that returns the full excerpt object when given an excerptId.