Initial commit of the interface between the View (React) and the Model (Database).
Probably took the coding a bit too far here because a lot of this logic will be handled by the database we choose. Until we have a database, we will likely never call "deleteArticle" for example. This is the Controller piece of the MVC that can't be implemented until we have a real Model.