A new CaptionView component that loads and converts the raw captions file into a specific HTML format.
All scripted lines become paragraphs with a data attribute added to the paragraph. The data attribute contains the start and end times of the caption that align with the video. The data attribute also contains a character offset for selecting text. 

Currently the caption automatically scrolls to the proper location when given a timer value in props.