Updated the functions to use our data structure to get the offset number instead of parsing the DOM.
Was able to do this everywhere except in getStartRangeOffsetTop, which needs to find the element by its offset in the DOM somehow. If this function can be re-engineered, then we can remove the data-offset attribute from the DOM altogether.

Also changed the way data-offset was calculated. Since we are using regular objects now instead of parsing one big HTML string, the offset doesn't need to include the content of the HTML tags. The new approach is simpler and more robust.

Still TODO, getCaptionNodeArray function.