Don't need data-start and data-end anymore, since we're using a map of word counts to timestamps.
Created a new type, ScrollData. The first number in the array is the word index, the second number is the timestamp for that word. This should be an object for readability, but since the value here get's added to a lot of DOM nodes it needs to be as concise as possible.