When the video timer changes, implement a sequence of binary searches to determine the proper scroll location. This is much more performant than the previous linear search method.
First find the index of the word in the captionWordMap. Then using this word index, find the index of the speaker in the speakerMap.

Using the speaker index we know which paragraph to scroll to. Within that paragraph, we can search for line breaks until we're reached the word index. The number of line breaks determines how far into this paragraph we need to scroll.