Reworked the Timeline to handle loading documents asynchronously.
Created a TimelineContainer component to fetch documents asynchronously. Since the server-side of this isn't ready yet, there is a setTimeout delay to mock the load time so that the loading component is displayed.

Created a TimelineLoadingView stateless functional component.

The Timeline component is now also stateless and renders based on its props which are determined by the TimelineContainer.

Created a temporary way to receive the foundation from the server, this is the object in newDb.ts, which will be deleted once the server is ready.

Had to replace all instances of the DocumentFact type with Foundation.DocumentFact throughout the project.