Using 3 different range sliders.
The first slider min value is 0, max value is video length. The range handles of this slider represent the zoom level min and max. 

The next two sliders min/max values are equal to the zoom level set by the first slider.

Slider 2 shows the text selection within the zoom level. Slider 3 shows the view window of the transcript within the zoom level.

The view window end range is hard coded to be 20 words greater than the start range. There is no easy way to get the real end range without significantly impacting performance, so hardcoded it for now.