Known issue: while loop executes faster than hiddenDiv.innerHTML reading/writing.
Get the index of the first word in the scrolling div. This is done by calculating the number of lines in the current paragraph based on its height and the value of the line-height. Then find the line number that is at the top of the div. This represents our "before" number line that starts at 0, goes to the number of lines in the paragraph, and the number we're interested in is the line number at the top of the view.

The "after" number line starts at the index of the first word in the paragraph and ends at the index of the last word in the paragraph. The number we are interested in is the index of the first word at the top of the view. Using numberLineTransform we can approximate this number, and then add or subtract words from the hidden div behind (z-index behind) the paragraph one word at a time. Once our div height matches the height we expect based on the previous calculations, we have the index of our first word.