Prevent the scroll event handler from doing any heavy lifting when the scroll is triggered by a change in props. Scroll event handler only needs to calculate the view range when the user scrolls the text.
May also consider not updating the current speaker name. On average this binary search takes 58ms, which is slower than the 60hz we need for a smooth animation.

(Measured with a small-ish sample size in Google Chrome on my box. Results will vary, but if my box can't handle it, the average user's box can't either)